refactor: reorg dtdt2 rules to prepare for v2 spec
