+++ cognite/neat/rules/importers/_dtdl2rules/spec/_v2_spec.py
@@ -1,0 +1,9 @@
+"""
+This is a pydantic validation implementation of the DTDL v2 spec.
+
+It is taken from https://github.com/Azure/opendigitaltwins-dtdl/blob/master/DTDL/v2/DTDL.v2.md
+"""
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    pass
