+++ tests/tests_unit/rules/test_importers/test_dtdl_importer.py
@@ -2,7 +2,7 @@ import pytest
 
 from cognite.neat.rules import validation
 from cognite.neat.rules.importers import DTDLImporter
-from cognite.neat.rules.importers._dtdl2rules._v3_spec import DTMI, Interface
+from cognite.neat.rules.importers._dtdl2rules.spec._v3_spec import DTMI, Interface
 from cognite.neat.rules.models._rules import InformationRules
 from cognite.neat.rules.models._rules.base import SchemaCompleteness
 from cognite.neat.rules.validation import IssueList
